Implement flat-fielding, and cropping
Add flat-fielding to the pipeline as a process object, as well as a
Cropper process object (the idea of the cropper being to crop out the
frilly edges, where bacteria will not be growing, and where our
edge-detection could be tripped up).

Include the ability to read/write to and from a .npy image for the flat
frame. Checks that the flat frame size matches the size of the images.